Add support for PWM override
This allow to set the PWM value of a shelf manually, adding a
Shelf.pwm_override attribute.
Overriding through Shelf.pwm is possible but will be lost on the next
PWM update.
It is also possible to set an expiration date for the PWM override
through the new Shelf.pwm_expire attribute, it uses datetime.datetime
objects to find out if the override has expired.
All the new attributes can be set to None do disable the override or
remove the expiration date.

Corresponding commands have been added to the fanctl cli.