Change util.terminate() to not call sys.exit()
util.terminate() was a non-returning function, which keept from raising
any exception after calling it and forced the program to quit.

Now, the function will return. It will just set the flag and call the
cleanup functions.

To actually exit the program, call util.sys_exit(), which is the wrapper
using the error message sent to util.terminate() if there was any.

Modified the main daemons function to raise an exception when a fatal
error occur, allowing the caller function to act accordingly.
The entry point functions will call the main daemon function, and will
then call util.sys_exit() to end the program with the correct return
code.